Test with 90 node(configure the limit of file descriptions) and refactor
