+++ core/tests/integration_test.rs
@@ -1,434 +1,14 @@
 mod common;
 use crate::common::VotationType;
 use crate::common::{
-    add_members_governance_request, create_genesis_event, create_governance_request,
-    create_register_event,
+    add_members_governance_request, add_providers, create_and_test_nodes,
+    create_and_test_nodes_with_trazabilty, create_event_governance, create_node,
+    create_nodes_and_connections, create_nodes_massive, get_request_with_votation, sign_events,
+    verify_copy_ledger,
 };
-use common::{generate_mc, McNodeData, NodeBuilder};
-use kore_base::request::{KoreRequest, RequestState};
-use kore_base::ApprovalState;
+use common::{McNodeData, NodeBuilder};
 use kore_base::EventRequest;
-use kore_base::{
-    keys::{Ed25519KeyPair, KeyGenerator, KeyPair},
-    DigestDerivator, DigestIdentifier,
-};
-use network::{NodeType, RoutingNode};
-use std::collections::HashSet;
-
-// Create node
-fn create_node(
-    addr: String,
-    node: NodeType,
-    vec_routing: Vec<RoutingNode>,
-    votation: VotationType,
-) -> (NodeBuilder, McNodeData) {
-    let key_pair = KeyPair::Ed25519(Ed25519KeyPair::new());
-    let mc_data_node1 = generate_mc(key_pair.clone());
-    let result = NodeBuilder::build(
-        node,
-        vec_routing,
-        vec![addr.to_string()],
-        votation,
-        key_pair.clone(),
-    )
-    .unwrap();
-    (result, mc_data_node1)
-}
-
-// Create governance and add keys
-async fn create_event_governance(node: NodeBuilder, mc_data_node: McNodeData) -> DigestIdentifier {
-    let public_key = node
-        .api
-        .add_keys(kore_base::KeyDerivator::Ed25519)
-        .await
-        .expect("MC creation failed");
-    let event_request = create_governance_request("".to_string(), public_key, "".to_string());
-    let request_id = sign_events(event_request, node.clone(), mc_data_node).await;
-    let subject_id = get_request_with_votation(request_id, node, VotationType::AlwaysAccept).await;
-    subject_id.subject_id.unwrap()
-}
-
-// Create genesis event
-async fn create_event_genesis(
-    node: NodeBuilder,
-    mc_data_node: McNodeData,
-    gobernance_id: DigestIdentifier,
-) -> DigestIdentifier {
-    let public_key = node
-        .api
-        .add_keys(kore_base::KeyDerivator::Ed25519)
-        .await
-        .expect("MC creation failed");
-    let event_request = create_genesis_event(
-        gobernance_id,
-        "".to_owned(),
-        public_key,
-        "traceability".to_owned(),
-    );
-    let request_id = sign_events(event_request, node.clone(), mc_data_node).await;
-    let subject_id = get_request_with_votation(request_id, node, VotationType::AlwaysAccept).await;
-    subject_id.subject_id.unwrap()
-}
-
-// Sign events and verify response
-async fn sign_events(
-    event: EventRequest,
-    node: NodeBuilder,
-    mc_data_node: McNodeData,
-) -> DigestIdentifier {
-    node.api
-        .external_request(mc_data_node.sign_event_request(&event, DigestDerivator::Blake3_512))
-        .await
-        .unwrap()
-}
-
-async fn get_request_with_votation(
-    request_id: DigestIdentifier,
-    node: NodeBuilder,
-    votation: VotationType,
-) -> KoreRequest {
-    let mut res;
-    loop {
-        res = node.api.get_request(request_id.clone()).await.unwrap();
-        match res.success {
-            Some(val) => match votation {
-                VotationType::AlwaysAccept => {
-                    assert_eq!(val, true);
-                    break;
-                }
-                VotationType::AlwaysReject => {
-                    assert_eq!(val, false);
-                    break;
-                }
-                VotationType::Normal => {
-                    if res.state == RequestState::Processing {
-                        // obtain the pending requests
-                        let res = node.api.get_pending_requests().await.unwrap();
-                        // aprove or reject the request
-                        for request in res.iter() {
-                            node.api
-                                .approval_request(request.id.clone(), true)
-                                .await
-                                .unwrap();
-                        }
-                    } else {
-                        assert!(val == true);
-                        break;
-                    }
-                }
-            },
-            None => {
-                tokio::time::sleep(std::time::Duration::from_millis(800)).await;
-            }
-        }
-        tokio::time::sleep(std::time::Duration::from_millis(800)).await;
-    }
-    res
-}
-
-// Authorize members to governance
-async fn add_providers(
-    node: NodeBuilder,
-    vec_providers: Vec<McNodeData>,
-    subject_id: DigestIdentifier,
-) {
-    let mut providers = HashSet::new();
-    for vec_provider in vec_providers.iter() {
-        providers.insert(vec_provider.get_controller_id());
-    }
-    node.api
-        .add_preauthorize_subject(&subject_id.clone(), &providers)
-        .await
-        .unwrap();
-}
-
-// Verify that the ledger is copied to all nodes
-async fn verify_copy_ledger(
-    vec_nodes: Vec<(NodeBuilder, McNodeData)>,
-    subject_id: DigestIdentifier,
-    sn: Option<u64>,
-) {
-    let mut pre_response = vec![];
-    for i in 0..vec_nodes.len() {
-        loop {
-            pre_response = vec_nodes[i]
-                .0
-                .api
-                .get_events(subject_id.clone(), None, None)
-                .await
-                .unwrap();
-            println!(" Response: {:?}", pre_response);
-            if pre_response.len() > 0
-                && pre_response[pre_response.len() - 1].content.sn == sn.unwrap_or(0)
-            {
-                break;
-            } else {
-                tokio::time::sleep(tokio::time::Duration::from_millis(300)).await;
-            }
-        }
-    }
-}
-
-// Deploy many nodes
-async fn create_nodes_massive(
-    size_bootstrap: usize,
-    size_addresable: usize,
-    size_ephermeral: usize,
-    initial_port: u32,
-) -> [Vec<(NodeBuilder, McNodeData)>; 3] {
-    let mut bootstrap_node: Vec<(NodeBuilder, McNodeData)> = vec![];
-    let mut routing_node: Vec<RoutingNode> = vec![];
-    let mut initial_port = initial_port;
-    for i in 0..size_bootstrap {
-        let addr = format!("/ip4/127.0.0.1/tcp/{:?}", initial_port);
-        let node_type = NodeType::Bootstrap;
-        if i == 0 {
-            bootstrap_node.push(create_node(
-                addr.clone(),
-                node_type,
-                vec![],
-                VotationType::AlwaysAccept,
-            ));
-        } else {
-            bootstrap_node.push(create_node(
-                addr.clone(),
-                node_type,
-                routing_node.clone(),
-                VotationType::AlwaysAccept,
-            ));
-        }
-        routing_node.push(RoutingNode {
-            peer_id: bootstrap_node[i].1.get_peer_id().to_string(),
-            address: vec![format!("/ip4/127.0.0.1/tcp/{:?}", initial_port)],
-        });
-        initial_port += 1;
-    }
-    let mut adressable_node = vec![];
-    for _i in 0..size_addresable {
-        let addr = format!("/ip4/127.0.0.1/tcp/{:?}", initial_port);
-        let node_type = NodeType::Addressable;
-        adressable_node.push(create_node(
-            addr.clone(),
-            node_type,
-            routing_node.clone(),
-            VotationType::AlwaysAccept,
-        ));
-        initial_port += 1;
-    }
-    let mut ephemeral_node = vec![];
-    for _i in 0..size_ephermeral {
-        let addr = format!("/ip4/127.0.0.1/tcp/{:?}", initial_port);
-        let node_type = NodeType::Ephemeral;
-        ephemeral_node.push(create_node(
-            addr.clone(),
-            node_type,
-            routing_node.clone(),
-            VotationType::AlwaysAccept,
-        ));
-        initial_port += 1;
-    }
-    [bootstrap_node, adressable_node, ephemeral_node]
-}
-
-// Create nodes with specific connections
-fn create_nodes_and_connections(
-    boostrap: Vec<Vec<usize>>,
-    addresable: Vec<Vec<usize>>,
-    ephemeral: Vec<Vec<usize>>,
-    initial_port: u32,
-) -> Vec<Vec<(NodeBuilder, McNodeData)>> {
-    let mut bootstrap_node: Vec<(NodeBuilder, McNodeData)> = vec![];
-    let mut addressable_node: Vec<(NodeBuilder, McNodeData)> = vec![];
-    let mut ephemeral_node: Vec<(NodeBuilder, McNodeData)> = vec![];
-    let mut routing_node: Vec<RoutingNode> = vec![];
-    let mut port = initial_port;
-    for i in 0..boostrap.len() {
-        if boostrap[i].len() == 0 {
-            bootstrap_node.push(create_node(
-                format!("/ip4/127.0.0.1/tcp/{:?}", port),
-                NodeType::Bootstrap,
-                vec![],
-                VotationType::AlwaysAccept,
-            ));
-            port += 1;
-        } else {
-            // iterate over the bootstrap nodes
-            for j in 0..boostrap[i].len() {
-                routing_node.push(RoutingNode {
-                    peer_id: bootstrap_node[boostrap[i][j]].1.get_peer_id().to_string(),
-                    address: vec![bootstrap_node[boostrap[i][j]].0.address[0].clone()],
-                });
-                // last item
-                if j == boostrap[i].len() - 1 {
-                    bootstrap_node.push(create_node(
-                        format!("/ip4/127.0.0.1/tcp/{:?}", port),
-                        NodeType::Bootstrap,
-                        routing_node.clone(),
-                        VotationType::AlwaysAccept,
-                    ));
-                    port += 1;
-                }
-            }
-            routing_node.clear();
-        }
-    }
-    port += 1;
-    routing_node.clear();
-
-    for i in 0..addresable.len() {
-        for j in 0..addresable[i].len() {
-            routing_node.push(RoutingNode {
-                peer_id: bootstrap_node[addresable[i][j]].1.get_peer_id().to_string(),
-                address: vec![bootstrap_node[addresable[i][j]].0.address[0].clone()],
-            });
-            if j == addresable[i].len() - 1 {
-                addressable_node.push(create_node(
-                    format!("/ip4/127.0.0.1/tcp/{:?}", port),
-                    NodeType::Addressable,
-                    routing_node.clone(),
-                    VotationType::AlwaysAccept,
-                ));
-                port += 1;
-            }
-        }
-        routing_node.clear();
-    }
-
-    port += 1;
-    routing_node.clear();
-
-    for i in 0..ephemeral.len() {
-        for j in 0..ephemeral[i].len() {
-            routing_node.push(RoutingNode {
-                peer_id: bootstrap_node[ephemeral[i][j]].1.get_peer_id().to_string(),
-                address: vec![bootstrap_node[ephemeral[i][j]].0.address[0].clone()],
-            });
-            if j == ephemeral[i].len() - 1 {
-                ephemeral_node.push(create_node(
-                    format!("/ip4/127.0.0.1/tcp/{:?}", port),
-                    NodeType::Ephemeral,
-                    routing_node.clone(),
-                    VotationType::AlwaysAccept,
-                ));
-                port += 1;
-            }
-        }
-        routing_node.clear();
-    }
-
-    vec![bootstrap_node, addressable_node, ephemeral_node]
-}
-
-// Test nodes with governance in diferent scenarios
-async fn create_and_test_nodes(vec_nodes: Vec<(NodeBuilder, McNodeData)>, index_governance: usize) {
-    // Create governance
-    let subject_id = create_event_governance(
-        vec_nodes[index_governance].0.clone(),
-        vec_nodes[index_governance].1.clone(),
-    )
-    .await;
-
-    // Add members to governance
-    let request_id = sign_events(
-        add_members_governance_request(
-            &vec_nodes,
-            subject_id.clone(),
-            index_governance,
-            "traceability".to_owned(),
-        ),
-        vec_nodes[index_governance].0.clone(),
-        vec_nodes[index_governance].1.clone(),
-    )
-    .await;
-    get_request_with_votation(
-        request_id,
-        vec_nodes[index_governance].0.clone(),
-        VotationType::AlwaysAccept,
-    )
-    .await;
-
-    // Add providers to governance
-    for i in 0..vec_nodes.len() {
-        add_providers(vec_nodes[i].0.clone(), vec![], subject_id.clone()).await
-    }
-    // Verify copy ledger
-    //verify_copy_ledger(vec_nodes.clone(), subject_id.clone(), None).await;
-}
-
-// Test nodes with governance in diferent scenarios
-async fn create_and_test_nodes_with_trazabilty(
-    vec_nodes: Vec<(NodeBuilder, McNodeData)>,
-    index_governance: usize,
-    index_emit: usize,
-) {
-    // Create governance
-    let subject_id = create_event_governance(
-        vec_nodes[index_governance].0.clone(),
-        vec_nodes[index_governance].1.clone(),
-    )
-    .await;
-    // Add members to governance
-    let request_id = sign_events(
-        add_members_governance_request(
-            &vec_nodes,
-            subject_id.clone(),
-            index_governance,
-            "traceability".to_owned(),
-        ),
-        vec_nodes[index_governance].0.clone(),
-        vec_nodes[index_governance].1.clone(),
-    )
-    .await;
-    get_request_with_votation(
-        request_id,
-        vec_nodes[index_governance].0.clone(),
-        VotationType::AlwaysAccept,
-    )
-    .await;
-    // Add providers to governance
-    for i in 0..vec_nodes.len() {
-        add_providers(vec_nodes[i].0.clone(), vec![], subject_id.clone()).await
-    }
-    // Verify copy ledger
-    //verify_copy_ledger(vec_nodes.clone(), subject_id.clone(), None).await;
-
-    // generate trazability with add keys and sign
-    let trazabilty_subject: DigestIdentifier = create_event_genesis(
-        vec_nodes[index_governance].0.clone(),
-        vec_nodes[index_governance].1.clone(),
-        subject_id.clone(),
-    )
-    .await;
-
-    let request_id = sign_events(
-        create_register_event(trazabilty_subject.clone()),
-        vec_nodes[index_emit].0.clone(),
-        vec_nodes[index_emit].1.clone(),
-    )
-    .await;
-
-    get_request_with_votation(
-        request_id,
-        vec_nodes[index_emit].0.clone(),
-        VotationType::AlwaysAccept,
-    )
-    .await;
-
-    // verify that the ledger is copied to all nodes
-    loop {
-        let response = vec_nodes[0]
-            .0
-            .api
-            .get_events(trazabilty_subject.clone(), None, None)
-            .await
-            .unwrap();
-        if response.len() > 0 {
-            break;
-        } else {
-            tokio::time::sleep(tokio::time::Duration::from_millis(300)).await;
-        }
-    }
-}
+use network::NodeType;
 
 #[cfg(test)]
 mod test {
